+++ src/test/java/chess/dao/user/UserDaoImplTest.java
@@ -1,6 +1,5 @@
-package chess.dao;
+package chess.dao.user;
 
-import chess.dao.user.UserDao;
 import chess.entity.UserEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
